+++ spec/unit/exceptions_spec.rb
@@ -3,12 +3,14 @@ require_relative './../spec_helper'
 # Tests for custom exception classes
 
 classes = [
+  OneviewSDK::OVError,
   OneviewSDK::ConnectionError,
   OneviewSDK::InvalidURL,
   OneviewSDK::InvalidClient,
   OneviewSDK::InvalidResource,
   OneviewSDK::IncompleteResource,
   OneviewSDK::MethodUnavailable,
+  OneviewSDK::UnsupportedVariant,
   OneviewSDK::UnsupportedVersion,
   OneviewSDK::InvalidRequest,
   OneviewSDK::BadRequest,
@@ -20,8 +22,18 @@ classes = [
 ]
 classes.each do |klass|
   RSpec.describe klass do
-    it 'exists and supports a message parameter' do
+    it 'behaves like the StandardError class' do
       expect { raise described_class, 'Msg' }.to raise_error(described_class, /Msg/)
+    end
+
+    it 'supports a message and data parameter' do
+      e = klass.new('Msg', key: :val)
+      expect(e.message).to eq('Msg')
+      expect(e.data).to eq(key: :val)
+    end
+
+    it 'has a shorthand raise! method' do
+      expect { described_class.raise! 'Msg', :data }.to raise_error(described_class, /Msg/)
     end
   end
 end